# [Enhancement] window function optimization, support rank (5)