--- a/be/src/exec/pipeline/sort/local_partition_topn_context.h
+++ b/be/src/exec/pipeline/sort/local_partition_topn_context.h
@@ -28,8 +28,8 @@ public:
     LocalPartitionTopnContext(const std::vector<TExpr>& t_partition_exprs, SortExecExprs& sort_exec_exprs,
                               std::vector<bool> is_asc_order, std::vector<bool> is_null_first,
                               const std::string& sort_keys, int64_t offset, int64_t partition_limit,
-                              const std::vector<OrderByType>& order_by_types, TupleDescriptor* materialized_tuple_desc,
-                              const RowDescriptor& parent_node_row_desc,
+                              const TTopNType::type topn_type, const std::vector<OrderByType>& order_by_types,
+                              TupleDescriptor* materialized_tuple_desc, const RowDescriptor& parent_node_row_desc,
                               const RowDescriptor& parent_node_child_row_desc);
 
     Status prepare(RuntimeState* state);
@@ -84,6 +84,7 @@ private:
     const std::string _sort_keys;
     int64_t _offset;
     int64_t _partition_limit;
+    const TTopNType::type _topn_type;
     const std::vector<OrderByType>& _order_by_types;
     TupleDescriptor* _materialized_tuple_desc;
     const RowDescriptor& _parent_node_row_desc;
@@ -99,7 +100,8 @@ public:
     LocalPartitionTopnContextFactory(const int32_t degree_of_parallelism, const std::vector<TExpr>& t_partition_exprs,
                                      SortExecExprs& sort_exec_exprs, std::vector<bool> is_asc_order,
                                      std::vector<bool> is_null_first, const std::string& sort_keys, int64_t offset,
-                                     int64_t partition_limit, const std::vector<OrderByType>& order_by_types,
+                                     int64_t partition_limit, const TTopNType::type topn_type,
+                                     const std::vector<OrderByType>& order_by_types,
                                      TupleDescriptor* materialized_tuple_desc,
                                      const RowDescriptor& parent_node_row_desc,
                                      const RowDescriptor& parent_node_child_row_desc);
@@ -118,6 +120,7 @@ private:
     const std::string _sort_keys;
     int64_t _offset;
     int64_t _partition_limit;
+    const TTopNType::type _topn_type;
     const std::vector<OrderByType>& _order_by_types;
     TupleDescriptor* _materialized_tuple_desc;
     const RowDescriptor& _parent_node_row_desc;