--- a/be/src/exec/pipeline/sort/sort_context.cpp
+++ b/be/src/exec/pipeline/sort/sort_context.cpp
@@ -23,7 +23,7 @@ ChunkPtr SortContext::pull_chunk() {
     }
     size_t required_rows = _state->chunk_size();
     required_rows = std::min<size_t>(required_rows, _total_rows);
-    if (_limit > 0) {
+    if (_limit > 0 && _topn_type == TTopNType::ROW_NUMBER) {
         required_rows = std::min<size_t>(required_rows, _limit);
     }
 
@@ -37,7 +37,12 @@ ChunkPtr SortContext::pull_chunk() {
 
 Status SortContext::_merge_inputs() {
     int64_t total_rows = _total_rows.load(std::memory_order_relaxed);
-    int64_t require_rows = ((_limit < 0) ? total_rows : std::min(_limit, total_rows));
+
+    // Keep all the data if topn type is RANK or DENSE_RANK
+    int64_t require_rows = total_rows;
+    if (_topn_type == TTopNType::ROW_NUMBER) {
+        require_rows = ((_limit < 0) ? total_rows : std::min(_limit, total_rows));
+    }
 
     std::vector<SortedRuns> partial_sorted_runs;
     for (int i = 0; i < _num_partition_sinkers; ++i) {
@@ -50,10 +55,12 @@ Status SortContext::_merge_inputs() {
     return Status::OK();
 }
 
-SortContextFactory::SortContextFactory(RuntimeState* state, bool is_merging, int64_t limit, int32_t num_right_sinkers,
+SortContextFactory::SortContextFactory(RuntimeState* state, const TTopNType::type topn_type, bool is_merging,
+                                       int64_t limit, int32_t num_right_sinkers,
                                        const std::vector<ExprContext*>& sort_exprs,
                                        const std::vector<bool>& is_asc_order, const std::vector<bool>& is_null_first)
         : _state(state),
+          _topn_type(topn_type),
           _is_merging(is_merging),
           _sort_contexts(is_merging ? 1 : num_right_sinkers),
           _limit(limit),
@@ -68,7 +75,7 @@ SortContextPtr SortContextFactory::create(int32_t idx) {
     DCHECK_LE(actual_idx, _sort_contexts.size());
     if (!_sort_contexts[actual_idx]) {
         _sort_contexts[actual_idx] =
-                std::make_shared<SortContext>(_state, _limit, num_sinkers, _sort_exprs, _sort_descs);
+                std::make_shared<SortContext>(_state, _topn_type, _limit, num_sinkers, _sort_exprs, _sort_descs);
     }
     return _sort_contexts[actual_idx];
 }