--- a/be/src/exec/vectorized/chunks_sorter.h
+++ b/be/src/exec/vectorized/chunks_sorter.h
@@ -123,9 +123,6 @@ public:
 
     virtual int64_t mem_usage() const = 0;
 
-    // For test only
-    void set_compare_strategy(CompareStrategy cmp) { _compare_strategy = cmp; }
-
 protected:
     size_t _get_number_of_order_by_columns() const { return _sort_exprs->size(); }
 
@@ -147,8 +144,6 @@ protected:
     RuntimeProfile::Counter* _output_timer = nullptr;
 
     std::atomic<bool> _is_sink_complete = false;
-
-    CompareStrategy _compare_strategy = ColumnWise;
 };
 
 } // namespace starrocks::vectorized