--- a/be/src/exec/vectorized/sorting/sort_permute.h
+++ b/be/src/exec/vectorized/sorting/sort_permute.h
@@ -10,6 +10,17 @@
 
 namespace starrocks::vectorized {
 
+enum RankType {
+    // Of this type, rank of [13, 13, 17, 17, 21, 25] is [1, 2, 3, 4, 5, 6]
+    RowNumber = 0,
+
+    // Of this type, rank of [13, 13, 17, 17, 21, 25] is [1, 1, 3, 3, 5, 6]
+    Rank = 1,
+
+    // Of this type, rank of [13, 13, 17, 17, 21, 25] is [1, 1, 2, 2, 3, 4]
+    DenseRank = 2,
+};
+
 enum CompareStrategy {
     Default = 0,
     RowWise = 1,
@@ -20,6 +31,7 @@ enum CompareStrategy {
 struct PermutationItem {
     uint32_t chunk_index;
     uint32_t index_in_chunk;
+    uint32_t rank_in_chunk;
 
     PermutationItem() = default;
     PermutationItem(uint32_t ci, uint32_t ii) : chunk_index(ci), index_in_chunk(ii) {}