--- a/be/src/exec/vectorized/sorting/sort_permute.h
+++ b/be/src/exec/vectorized/sorting/sort_permute.h
@@ -10,13 +10,6 @@
 
 namespace starrocks::vectorized {
 
-enum CompareStrategy {
-    Default = 0,
-    RowWise = 1,
-    ColumnWise = 2,
-    ColumnInc = 3,
-};
-
 struct PermutationItem {
     uint32_t chunk_index;
     uint32_t index_in_chunk;