… be/src/exec/vectorized/sorting/sorting.h
… be/src/exec/vectorized/topn_node.cpp
… be/test/exec/vectorized/chunks_sorter_bench_test.cpp
… be/test/exec/vectorized/chunks_sorter_test.cpp
… fe/fe-core/src/main/java/com/starrocks/planner/ExchangeNode.java
… fe/fe-core/src/main/java/com/starrocks/planner/SortNode.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/base/SortProperty.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/TopNType.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalTopNOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalTopNOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/AddDecodeNodeForDictStringRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/ExchangeSortToMergeRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/TopNImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateWindowRankRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
… gensrc/thrift/PlanNodes.thrift