@wanpengfei-git 2022-05-20
#### [FE PR Coverage check]

:heart_eyes: **pass** : 168 / 185 (90.81%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/analysis/SortInfo.java|2|3|66.67%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/TopNType.java|9|13|69.23%|
|:large_blue_circle:|com/starrocks/sql/optimizer/transformer/QueryTransformer.java|11|14|78.57%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/logical/LogicalTopNOperator.java|25|31|80.65%|
|:large_blue_circle:|com/starrocks/planner/SortNode.java|24|26|92.31%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateWindowRankRule.java|56|57|98.25%|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/Operator.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/TopNImplementationRule.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|18|18|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rewrite/ExchangeSortToMergeRule.java|4|4|100.00%|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|8|8|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/RuleType.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalTopNOperator.java|4|4|100.00%|
|:large_blue_circle:|com/starrocks/analysis/QueryStmt.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rewrite/AddDecodeNodeForDictStringRule.java|1|1|100.00%|