@wanpengfei-git 2022-05-25
#### [FE PR Coverage check]

:heart_eyes: **pass** : 50 / 54 (92.59%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/TopNType.java|9|13|69.23%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/logical/LogicalTopNOperator.java|11|11|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateWindowRankRule.java|5|5|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/TopNImplementationRule.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/base/SortProperty.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|8|8|100.00%|
|:large_blue_circle:|com/starrocks/planner/SortNode.java|7|7|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rewrite/ExchangeSortToMergeRule.java|3|3|100.00%|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalTopNOperator.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rewrite/AddDecodeNodeForDictStringRule.java|1|1|100.00%|